Refactor code to use an application struct
